A start
Here, PORTB4 goes up throughout the first byte, but for subsequent
bytes it goes up only at the end of the byte, and only for a few us.

Subsequent bytes are received and processed.

I theorise that INT0 is being triggered as soon as it is
re-enabled (which is after the end of the byte) and perhaps
the receive state machine is looping in state zero.